+++ apps/web/playwright/organization/organization-invitation.e2e.ts
@@ -106,7 +106,7 @@ test.describe("Organization", () => {
         const invitedUserEmail = users.trackEmail({ username: "rick", domain: "domain.com" });
         // '-domain' because the email doesn't match orgAutoAcceptEmail
         const usernameDerivedFromEmail = `${invitedUserEmail.split("@")[0]}-domain`;
-        await inviteAnEmail(page, invitedUserEmail);
+        await inviteAnEmail(page, invitedUserEmail, true);
         await expectUserToBeAMemberOfTeam({
           page,
           teamId: team.id,
@@ -164,7 +164,7 @@ test.describe("Organization", () => {
 
       await test.step("By invite link", async () => {
         await page.goto(`/settings/teams/${team.id}/members`);
-        const inviteLink = await copyInviteLink(page);
+        const inviteLink = await copyInviteLink(page, true);
         const email = users.trackEmail({ username: "rick", domain: "domain.com" });
         // '-domain' because the email doesn't match orgAutoAcceptEmail
         const usernameDerivedFromEmail = `${email.split("@")[0]}-domain`;
@@ -297,6 +297,9 @@ test.describe("Organization", () => {
       await test.step("Signing up with the previous username of the migrated user - shouldn't be allowed", async () => {
         await page.goto("/signup");
         await expect(page.locator("text=Create your account")).toBeVisible();
+        await expect(page.locator('[data-testid="continue-with-email-button"]')).toBeVisible();
+        await page.locator('[data-testid="continue-with-email-button"]').click();
+        await expect(page.locator('[data-testid="signup-submit-button"]')).toBeVisible();
 
         // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
         await page.locator('input[name="username"]').fill(existingUser.username!);
@@ -330,7 +333,7 @@ test.describe("Organization", () => {
         await page.goto(`/settings/teams/${team.id}/members`);
         const invitedUserEmail = users.trackEmail({ username: "rick", domain: "example.com" });
         const usernameDerivedFromEmail = invitedUserEmail.split("@")[0];
-        await inviteAnEmail(page, invitedUserEmail);
+        await inviteAnEmail(page, invitedUserEmail, true);
         await expectUserToBeAMemberOfTeam({
           page,
           teamId: team.id,
@@ -388,7 +391,7 @@ test.describe("Organization", () => {
       await test.step("By invite link", async () => {
         await page.goto(`/settings/teams/${team.id}/members`);
 
-        const inviteLink = await copyInviteLink(page);
+        const inviteLink = await copyInviteLink(page, true);
         const email = users.trackEmail({ username: "rick", domain: "example.com" });
         // '-domain' because the email doesn't match orgAutoAcceptEmail
         const usernameDerivedFromEmail = `${email.split("@")[0]}`;
@@ -478,8 +481,12 @@ export async function signupFromEmailInviteLink({
   await signupPage.close();
 }
 
-async function inviteAnEmail(page: Page, invitedUserEmail: string) {
-  await page.locator('button:text("Add")').click();
+async function inviteAnEmail(page: Page, invitedUserEmail: string, teamPage?: boolean) {
+  if (teamPage) {
+    await page.getByTestId("new-member-button").click();
+  } else {
+    await page.getByTestId("new-organization-member-button").click();
+  }
   await page.locator('input[name="inviteUser"]').fill(invitedUserEmail);
   const submitPromise = page.waitForResponse("/api/trpc/teams/inviteMember?batch=1");
   await page.locator('button:text("Send invite")').click();
@@ -555,8 +562,12 @@ function assertInviteLink(inviteLink: string | null | undefined): asserts invite
   if (!inviteLink) throw new Error("Invite link not found");
 }
 
-async function copyInviteLink(page: Page) {
-  await page.locator('button:text("Add")').click();
+async function copyInviteLink(page: Page, teamPage?: boolean) {
+  if (teamPage) {
+    await page.getByTestId("new-member-button").click();
+  } else {
+    await page.getByTestId("new-organization-member-button").click();
+  }
   const inviteLink = await getInviteLink(page);
   return inviteLink;
 }