+++ apps/web/test/lib/handleChildrenEventTypes.test.ts
@@ -40,8 +40,6 @@ describe("handleChildrenEventTypes", () => {
         children: [],
         updatedEventType: { schedulingType: null, slug: "something" },
         currentUserId: 1,
-        hashedLink: undefined,
-        connectedLink: null,
         prisma: prismaMock,
         profileId: null,
         updatedValues: {},
@@ -64,8 +62,6 @@ describe("handleChildrenEventTypes", () => {
         children: [],
         updatedEventType: { schedulingType: "MANAGED", slug: "something" },
         currentUserId: 1,
-        hashedLink: undefined,
-        connectedLink: null,
         prisma: prismaMock,
         profileId: null,
         updatedValues: {},
@@ -91,8 +87,6 @@ describe("handleChildrenEventTypes", () => {
         children: [],
         updatedEventType: { schedulingType: "MANAGED", slug: "something" },
         currentUserId: 1,
-        hashedLink: undefined,
-        connectedLink: null,
         prisma: prismaMock,
         profileId: null,
         updatedValues: {},
@@ -131,8 +125,6 @@ describe("handleChildrenEventTypes", () => {
         children: [{ hidden: false, owner: { id: 4, name: "", email: "", eventTypeSlugs: [] } }],
         updatedEventType: { schedulingType: "MANAGED", slug: "something" },
         currentUserId: 1,
-        hashedLink: undefined,
-        connectedLink: null,
         prisma: prismaMock,
         profileId: null,
         updatedValues: {},
@@ -186,8 +178,6 @@ describe("handleChildrenEventTypes", () => {
         children: [{ hidden: false, owner: { id: 4, name: "", email: "", eventTypeSlugs: [] } }],
         updatedEventType: { schedulingType: "MANAGED", slug: "something" },
         currentUserId: 1,
-        hashedLink: "somestring",
-        connectedLink: null,
         prisma: prismaMock,
         profileId: null,
         updatedValues: {
@@ -202,8 +192,10 @@ describe("handleChildrenEventTypes", () => {
           scheduleId: null,
           lockTimeZoneToggleOnBookingPage: false,
           requiresBookerEmailVerification: false,
+          hashedLink: {
+            deleteMany: {},
+          },
           instantMeetingScheduleId: undefined,
-          hashedLink: { create: { link: expect.any(String) } },
         },
         where: {
           userId_parentId: {
@@ -226,8 +218,6 @@ describe("handleChildrenEventTypes", () => {
         children: [],
         updatedEventType: { schedulingType: "MANAGED", slug: "something" },
         currentUserId: 1,
-        hashedLink: undefined,
-        connectedLink: null,
         prisma: prismaMock,
         profileId: null,
         updatedValues: {},
@@ -252,8 +242,6 @@ describe("handleChildrenEventTypes", () => {
         ],
         updatedEventType: { schedulingType: "MANAGED", slug: "something" },
         currentUserId: 1,
-        hashedLink: undefined,
-        connectedLink: null,
         prisma: prismaMock,
         profileId: null,
         updatedValues: {},
@@ -293,8 +281,6 @@ describe("handleChildrenEventTypes", () => {
         children: [{ hidden: false, owner: { id: 4, name: "", email: "", eventTypeSlugs: ["something"] } }],
         updatedEventType: { schedulingType: "MANAGED", slug: "something" },
         currentUserId: 1,
-        hashedLink: undefined,
-        connectedLink: null,
         prisma: prismaMock,
         profileId: null,
         updatedValues: {},
@@ -308,12 +294,12 @@ describe("handleChildrenEventTypes", () => {
           durationLimits: undefined,
           recurringEvent: undefined,
           eventTypeColor: undefined,
-          hashedLink: undefined,
           instantMeetingScheduleId: undefined,
           lockTimeZoneToggleOnBookingPage: false,
           requiresBookerEmailVerification: false,
           userId: 4,
           workflows: undefined,
+          hashedLink: undefined,
         },
       });
       expect(result.newUserIds).toEqual([4]);
@@ -349,8 +335,6 @@ describe("handleChildrenEventTypes", () => {
         children: [{ hidden: false, owner: { id: 4, name: "", email: "", eventTypeSlugs: ["something"] } }],
         updatedEventType: { schedulingType: "MANAGED", slug: "something" },
         currentUserId: 1,
-        hashedLink: undefined,
-        connectedLink: null,
         prisma: prismaMock,
         profileId: null,
         updatedValues: {
@@ -362,6 +346,9 @@ describe("handleChildrenEventTypes", () => {
         data: {
           ...rest,
           locations: [],
+          hashedLink: {
+            deleteMany: {},
+          },
           lockTimeZoneToggleOnBookingPage: false,
           requiresBookerEmailVerification: false,
           instantMeetingScheduleId: undefined,
@@ -416,8 +403,6 @@ describe("handleChildrenEventTypes", () => {
         ],
         updatedEventType: { schedulingType: "MANAGED", slug: "something" },
         currentUserId: 1,
-        hashedLink: undefined,
-        connectedLink: null,
         prisma: prismaMock,
         profileId: null,
         updatedValues: {},
@@ -429,7 +414,6 @@ describe("handleChildrenEventTypes", () => {
           durationLimits: undefined,
           recurringEvent: undefined,
           eventTypeColor: undefined,
-          hashedLink: undefined,
           instantMeetingScheduleId: undefined,
           locations: [],
           lockTimeZoneToggleOnBookingPage: false,
@@ -446,6 +430,7 @@ describe("handleChildrenEventTypes", () => {
           workflows: {
             create: [{ workflowId: 11 }],
           },
+          hashedLink: undefined,
         },
       });
       const { profileId, ...rest } = evType;
@@ -456,7 +441,9 @@ describe("handleChildrenEventTypes", () => {
           locations: [],
           lockTimeZoneToggleOnBookingPage: false,
           requiresBookerEmailVerification: false,
-          hashedLink: undefined,
+          hashedLink: {
+            deleteMany: {},
+          },
           instantMeetingScheduleId: undefined,
         },
         where: {