+++ packages/app-store/googlecalendar/lib/CalendarService.ts
@@ -16,6 +16,7 @@ import {
   CREDENTIAL_SYNC_SECRET_HEADER_NAME,
 } from "@calcom/lib/constants";
 import { formatCalEvent } from "@calcom/lib/formatCalendarEvent";
+import { getAllCalendars } from "@calcom/lib/google";
 import logger from "@calcom/lib/logger";
 import { safeStringify } from "@calcom/lib/safeStringify";
 import prisma from "@calcom/prisma";
@@ -256,6 +257,9 @@ export default class GoogleCalendarService implements Calendar {
         : true,
       iCalUID: formattedCalEvent.iCalUID,
     };
+    if (calEventRaw.hideCalendarEventDetails) {
+      payload.visibility = "private";
+    }
 
     if (formattedCalEvent.location) {
       payload["location"] = getLocation(formattedCalEvent);
@@ -606,9 +610,9 @@ export default class GoogleCalendarService implements Calendar {
     }
     async function getCalIds() {
       if (selectedCalendarIds.length !== 0) return selectedCalendarIds;
-      const cals = await calendar.calendarList.list({ fields: "items(id)" });
-      if (!cals.data.items) return [];
-      return cals.data.items.reduce((c, cal) => (cal.id ? [...c, cal.id] : c), [] as string[]);
+      const cals = await getAllCalendars(calendar, ["id"]);
+      if (!cals.length) return [];
+      return cals.reduce((c, cal) => (cal.id ? [...c, cal.id] : c), [] as string[]);
     }
 
     try {
@@ -666,9 +670,13 @@ export default class GoogleCalendarService implements Calendar {
     try {
       const { json: cals } = await this.oAuthManagerInstance.request(
         async () =>
-          new AxiosLikeResponseToFetchResponse(
-            await calendar.calendarList.list({ fields: "items(id,summary,primary,accessRole)" })
-          )
+          new AxiosLikeResponseToFetchResponse({
+            status: 200,
+            statusText: "OK",
+            data: {
+              items: await getAllCalendars(calendar),
+            },
+          })
       );
 
       if (!cals.items) return [];