+++ packages/app-store/routing-forms/lib/createFallbackRoute.ts
@@ -1,6 +1,7 @@
 import { Utils as QbUtils } from "react-awesome-query-builder";
 
-import type { GlobalRoute, SerializableRoute } from "../types/types";
+import type { GlobalRoute, LocalRoute, SerializableRoute } from "../types/types";
+import { RouteActionType } from "../zod";
 
 export const createFallbackRoute = (): Exclude<SerializableRoute, GlobalRoute> => {
   const uuid = QbUtils.uuid();
@@ -8,9 +9,9 @@ export const createFallbackRoute = (): Exclude<SerializableRoute, GlobalRoute> =
     id: uuid,
     isFallback: true,
     action: {
-      type: "customPageMessage",
+      type: RouteActionType.CustomPageMessage,
       value: "Thank you for your interest! We will be in touch soon.",
     },
-    queryValue: { id: uuid, type: "group" },
+    queryValue: { id: uuid, type: "group" } as LocalRoute["queryValue"],
   };
 };