+++ packages/lib/server/queries/teams/index.ts
@@ -261,7 +261,7 @@ export async function getTeamWithMembers(args: {
   };
 }
 
-export async function getMinimalTeam(args: {
+export async function getTeamWithoutMembers(args: {
   id?: number;
   slug?: string;
   userId?: number;
@@ -299,6 +299,7 @@ export async function getMinimalTeam(args: {
       isPrivate: true,
       metadata: true,
       bookingLimits: true,
+      includeManagedEventsInLimits: true,
       parent: {
         select: {
           id: true,
@@ -418,7 +419,9 @@ export async function updateNewTeamMemberEventTypes(userId: number, teamId: numb
     },
   });
 
-  const allManagedEventTypePropsZod = _EventTypeModel.pick(allManagedEventTypeProps);
+  const allManagedEventTypePropsZod = _EventTypeModel.pick(allManagedEventTypeProps).extend({
+    bookingFields: _EventTypeModel.shape.bookingFields.nullish(),
+  });
 
   eventTypesToAdd.length > 0 &&
     (await prisma.$transaction(
@@ -453,7 +456,7 @@ export async function updateNewTeamMemberEventTypes(userId: number, teamId: numb
               users: {
                 connect: [{ id: userId }],
               },
-              parentId: eventType.parentId,
+              parentId: eventType.id,
               hidden: false,
               workflows: currentWorkflowIds && {
                 create: currentWorkflowIds.map((wfId) => ({ workflowId: wfId })),