+++ packages/prisma/migrations/20241010070020_add_active/migration.sql
@@ -1,0 +1,2 @@
+-- CreateIndex
+CREATE INDEX "Webhook_active_idx" ON "Webhook"("active");