+++ packages/types/CrmService.d.ts
@@ -25,9 +25,22 @@ export interface CrmEvent {
 }
 
 export interface CRM {
-  createEvent: (event: CalendarEvent, contacts: Contact[]) => Promise<CrmEvent>;
+  createEvent: (event: CalendarEvent, contacts: Contact[]) => Promise<CrmEvent | undefined>;
   updateEvent: (uid: string, event: CalendarEvent) => Promise<CrmEvent>;
   deleteEvent: (uid: string) => Promise<void>;
-  getContacts: (emails: string | string[], includeOwner?: boolean) => Promise<Contact[]>;
+  getContacts: ({
+    emails,
+    includeOwner,
+    forRoundRobinSkip,
+  }: {
+    emails: string | string[];
+    includeOwner?: boolean;
+    forRoundRobinSkip?: boolean;
+  }) => Promise<Contact[]>;
   createContacts: (contactsToCreate: ContactCreateInput[], organizerEmail?: string) => Promise<Contact[]>;
+  getAppOptions: () => any;
+  handleAttendeeNoShow?: (
+    bookingUid: string,
+    attendees: { email: string; noShow: boolean }[]
+  ) => Promise<void>;
 }