+++ packages/platform/libraries/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.51
+Released to support PR https://github.com/calcom/cal.com/pull/17224 which enabled attendee specified location during booking.
+
+## 0.0.41
+Released to support handle cancel booking passing oauth client id to webhooks.
+
 ## 0.0.38
 #### Feature: Added Support for AdvancedTab Event-Type Attributes in API
 - **Booker Layouts**: